requirements.txt
bcrypt==3.1.4
blinker==1.4
certifi==2016.2.28
cffi
click==8.1.7
flask
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2
itsdangerous==0.24
Jinja2==3.1
MarkupSafe
Pillow
pycparser==2.18
six==1.11.0
SQLAlchemy==2.0.29
Werkzeug==3.0
WTForms==3.1
authlib
requests
cryptography
